update these code

+++ iteastyle-dal/src/main/java/com/aimeow/iteastyle/dao/PostDAO.java
@@ -10,7 +10,7 @@ public interface PostDAO {
         PostQuery query) throws Exception;
     PostDO queryPostById(
         PostQuery query) throws Exception;
-    Integer countPosts(
+    Long countPosts(
         PostQuery query) throws Exception;
     Boolean createPost(
         PostDO postDO) throws Exception;
